requirements.txt
torch
transformers
accelerate
bitsandbytes
sentence-transformers
faiss-cpu
pdfplumber
nltk
rouge-score
bert-score
scikit-learn
tqdm